+++ src/Dfe.Analytics.EFCore/Description/AnnotationKeys.cs
@@ -4,6 +4,7 @@ internal static class AnnotationKeys
 {
     private const string BaseKeyName = "DfeAnalytics";
 
+    internal const string DatabaseSyncMetadata = BaseKeyName + nameof(DatabaseSyncMetadata);
     internal const string TableAnalyticsSyncMetadata = BaseKeyName + nameof(TableAnalyticsSyncMetadata);
     internal const string ColumnAnalyticsSyncMetadata = BaseKeyName + nameof(ColumnAnalyticsSyncMetadata);